requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.1
discord.py==2.3.2
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
multidict==6.0.4
Pillow==10.1.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
SQLAlchemy==2.0.22
typing_extensions==4.8.0
yarl==1.9.2
